extracted player and profile specific storage into their own classes in a new packet
