+++ src/main/java/at/hannibal2/skyhanni/features/minion/MinionFeatures.kt
@@ -2,7 +2,7 @@ package at.hannibal2.skyhanni.features.minion
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.config.ConfigUpdaterMigrator
-import at.hannibal2.skyhanni.config.Storage
+import at.hannibal2.skyhanni.config.storage.ProfileSpecificStorage
 import at.hannibal2.skyhanni.data.ClickType
 import at.hannibal2.skyhanni.data.IslandType
 import at.hannibal2.skyhanni.data.ProfileStorageData
@@ -176,7 +176,7 @@ class MinionFeatures {
         val name = getMinionName(openInventory)
         if (!minions.contains(entity) && LorenzUtils.skyBlockIsland != IslandType.HUB) {
             MinionFeatures.minions = minions.editCopy {
-                this[entity] = Storage.ProfileSpecific.MinionConfig().apply {
+                this[entity] = ProfileSpecificStorage.MinionConfig().apply {
                     displayName = name
                     lastClicked = 0
                 }
@@ -288,7 +288,7 @@ class MinionFeatures {
         }
         if (message.startsWith("§bYou placed a minion!") && newMinion != null) {
             minions = minions?.editCopy {
-                this[newMinion!!] = Storage.ProfileSpecific.MinionConfig().apply {
+                this[newMinion!!] = ProfileSpecificStorage.MinionConfig().apply {
                     displayName = newMinionName
                     lastClicked = 0
                 }
@@ -376,7 +376,7 @@ class MinionFeatures {
         var minionInventoryOpen = false
         var minionStorageInventoryOpen = false
 
-        private var minions: Map<LorenzVec, Storage.ProfileSpecific.MinionConfig>?
+        private var minions: Map<LorenzVec, ProfileSpecificStorage.MinionConfig>?
             get() {
                 return ProfileStorageData.profileSpecific?.minions
             }
