+++ src/main/java/at/hannibal2/skyhanni/features/slayer/SlayerProfitTracker.kt
@@ -2,7 +2,7 @@ package at.hannibal2.skyhanni.features.slayer
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.config.ConfigUpdaterMigrator
-import at.hannibal2.skyhanni.config.Storage
+import at.hannibal2.skyhanni.config.storage.ProfileSpecificStorage
 import at.hannibal2.skyhanni.data.SlayerAPI
 import at.hannibal2.skyhanni.data.jsonobjects.repo.SlayerProfitTrackerItemsJson
 import at.hannibal2.skyhanni.events.GuiRenderEvent
@@ -108,7 +108,7 @@ object SlayerProfitTracker {
         if (itemLogCategory == "") return null
 
         return trackers.getOrPut(itemLogCategory) {
-            val getStorage: (Storage.ProfileSpecific) -> Data = {
+            val getStorage: (ProfileSpecificStorage) -> Data = {
                 it.slayerProfitData.getOrPut(
                     itemLogCategory
                 ) { Data() }
