+++ src/main/java/at/hannibal2/skyhanni/test/SkyHanniConfigSearchResetCommand.kt
@@ -224,7 +224,7 @@ object SkyHanniConfigSearchResetCommand {
                 if (!classFilter(className)) continue
                 val objectName = obj.getObjectName()
                 if (obj !is Runnable && objectName.startsWith(className) && (objectName.startsWith("at.hannibal2.skyhanni.config.features.") ||
-                        objectName.startsWith("at.hannibal2.skyhanni.config.Storage"))
+                        objectName.startsWith("at.hannibal2.skyhanni.config.storage.Storage"))
                 ) {
                     "<category>"
                 } else if (onlyValue) {
